Refactor: Move TableReportPage functionality into a standalone TableReport component
- Migrated logic and UI from TableReportPage to a new TableReport component for better reusability.
- Removed the redundant TableReportPage.jsx file.
- Updated references in App.jsx and ReportTabs.jsx to use the new TableReport component.
